Replace usage of `COMMONS[...]` with `getcommon()` from `SuiteSparse` so that the tests can run.
[SuiteSparse](https://github.com/JuliaSparse/SuiteSparse.jl) has been
merged into
[SparseArrays](https://github.com/JuliaSparse/SparseArrays.jl) and now
only contains a wrapper to the latter package. However, since the
version compatibility is not restricted, the current calls in
`from_cholmod.jl` does not work anymore without some small changes.

(Initially I tried to drop the dependency on `SuiteSparse`, but it turns
out that `SparseArrays` from older Julia versions are not aware of those
changes either)